Move list filtering base tests into tests.utils
