+++ keystone_api/tests/allocations/test_allocations.py
@@ -3,8 +3,9 @@
 from rest_framework import status
 from rest_framework.test import APITestCase
 
+from apps.allocations.models import Allocation
 from apps.users.models import User
-from tests.utils import CustomAsserts
+from tests.utils import CustomAsserts, ListEndpointFilteringTests
 
 
 class EndpointPermissions(APITestCase, CustomAsserts):
@@ -75,3 +76,9 @@ class EndpointPermissions(APITestCase, CustomAsserts):
             trace=status.HTTP_405_METHOD_NOT_ALLOWED,
             post_body={'requested': 1000, 'cluster': 1, 'request': 1}
         )
+
+class RecordFiltering(ListEndpointFilteringTests, APITestCase):
+    """Test the filtering of returned records based on user team membership."""
+
+    endpoint = '/allocations/allocations/'
+    model = Allocation
